Improved the way we handle e2e tests
Expected paths for files after generation are now placed in different files whether they're common or app type specific ( nativescript )
Factorized getExpectedFilesPaths into common.utils.ts
Added a plugin-name.ts file in each e2e app to hold the plugin's name as a const for all of its e2e tests
From now on, we'll go on a "one test file per generator" policy
