+++ e2e/nativescript-e2e/tests/app.spec.ts
@@ -11,45 +11,25 @@ import {
  * Internal imports
  */
 import {
+    getExpectedFilesPaths,
+    runGeneratorForPlugin
+} from '../../common.utils';
+import {
     APP_RESOURCES_FILES,
-    HOOKS_FILES,
-    ROOT_FILES,
-    SRC_FILES
+    HOOKS_FILES
+} from './apps-structures/common';
+import {
+    ANGULAR_ROOT_FILES,
+    ANGULAR_SRC_FILES
 } from './apps-structures/angular';
-import { runGeneratorForPlugin } from '../../common.utils';
+import { PLUGIN_NAME } from './plugin-name';
 
 /**
  * TypeScript entities and constants
  */
-const pluginName = 'nativescript';
+const generatorName = 'app';
 
-describe('The nativescript plugin, with the app generator,', () => {
-    const generatorName = 'app';
-
-    /**
-     * Helper that completes the expected files paths with
-     * the newly generated app name
-     *
-     * @param filesPaths : string[], the paths of the files expected after
-     * generation is complete, relative to the root of that new app
-     * @param appName : string, this new app's name
-     * @return the paths of the files expected after
-     * generation is complete, relative to the root of the Nx workspace
-     */
-    const getExpectedFilesPaths: (
-        filesPaths: string[],
-        appName: string
-    ) => string[] =
-        (
-            filesPaths: string[],
-            appName: string
-        ) => {
-            return filesPaths.map(
-                (filePath: string) => {
-                    return `apps/${appName}/${filePath}`;
-                }
-            );
-        };
+describe(`The ${PLUGIN_NAME} plugin, with the ${generatorName} generator,`, () => {
 
     /**
      * Helper that tests if the generated application is able to build
@@ -58,18 +38,18 @@ describe('The nativescript plugin, with the app generator,', () => {
      * @param platform : 'ios' | 'android', the mobile platform you want to test
      * the build for
      */
-    const testBuildForPlatform: (platform: 'ios' | 'android') => Promise<void> = async (platform: 'ios' | 'android') => {
-        const appName: string = await runGeneratorForPlugin(pluginName, generatorName);
+    async function testBuildForPlatform(platform: 'ios' | 'android'): Promise<void> {
+        const appName: string = await runGeneratorForPlugin(PLUGIN_NAME, generatorName);
         const result: { stdout: string; stderr: string } = await runNxCommandAsync(`build ${appName} -c ${platform}`);
         expect(result.stdout).toContain('Project successfully built');
-    };
+    }
 
     it('should create a Nativescript+Angular app by default', async () => {
-        const appName: string = await runGeneratorForPlugin(pluginName, generatorName);
+        const appName: string = await runGeneratorForPlugin(PLUGIN_NAME, generatorName);
 
         // Computing expected files paths
-        const rootFiles: string[] = getExpectedFilesPaths(ROOT_FILES, appName);
-        const srcFiles: string[] = getExpectedFilesPaths(SRC_FILES, appName);
+        const rootFiles: string[] = getExpectedFilesPaths(ANGULAR_ROOT_FILES, appName);
+        const srcFiles: string[] = getExpectedFilesPaths(ANGULAR_SRC_FILES, appName);
         const hooksFiles: string[] = getExpectedFilesPaths(HOOKS_FILES, appName);
         const appResourcesFiles: string[] = getExpectedFilesPaths(APP_RESOURCES_FILES, appName);
 
@@ -95,7 +75,7 @@ describe('The nativescript plugin, with the app generator,', () => {
     describe('and the --directory option,', () => {
         it('should create src in the specified directory', async () => {
             const directory: string = 'subdir';
-            const appName: string = await runGeneratorForPlugin(pluginName, generatorName, `--directory ${directory}`);
+            const appName: string = await runGeneratorForPlugin(PLUGIN_NAME, generatorName, `--directory ${directory}`);
             expect(() =>
                        checkFilesExist(`apps/${directory}/${appName}/src/main.ts`)
             ).not.toThrow();
@@ -105,7 +85,7 @@ describe('The nativescript plugin, with the app generator,', () => {
     describe('and the --tags option', () => {
         it('should add tags to nx.json', async () => {
             const tags: string[] = ['e2etag', 'e2ePackage'];
-            const appName: string = await runGeneratorForPlugin(pluginName, generatorName, `--tags ${tags.join(',')}`);
+            const appName: string = await runGeneratorForPlugin(PLUGIN_NAME, generatorName, `--tags ${tags.join(',')}`);
             const nxJson: any = readJson('nx.json');
             expect(nxJson.projects[appName].tags).toEqual(['e2etag', 'e2ePackage']);
         });
