+++ e2e/nest-sls-e2e/tests/plugin-name.ts
@@ -1,0 +1,1 @@
+export const PLUGIN_NAME: string = 'nest-sls';
